requirements.txt
Flask==2.0.1
python-dotenv==0.19.2
teleflask==2.0.0.dev23
